+++ uui/src/lib/List/List.types.d.ts
@@ -1,4 +1,5 @@
-import type { BoxProps } from '../Box/box.types';
-export interface ListProps<T extends EventTarget> extends BoxProps<T> {
+import type { PaperProps } from '../Paper/Paper.types';
+
+export interface ListProps<T extends EventTarget> extends PaperProps<T> {
   disablePadding?: boolean;
 }