+++ uui/src/lib/List/ListItem.types.d.ts
@@ -1,11 +1,7 @@
 import type { BoxProps } from '../Box/box.types';
 
-export type ListItemSlotProps = Record<string, boolean | undefined>;
-
 export interface ListItemProps<T extends EventTarget> extends BoxProps<T> {
-  text: string;
-  secondaryText?: string;
-  textColor?: string;
+  textColor?: ThemeColors;
   selected?: boolean;
   divider?: boolean;
   disabled?: boolean;