+++ website/src/routes/list-item/leading-icon-list.svelte
@@ -1,22 +1,24 @@
-<script>
-  import { List, ListItem, ListItemAdornment } from '@uui';
+<script lang="ts">
+  import { List, ListItem, ListItemAdornment, Typography } from '@uui';
+
+  let selectedIndex = 0;
+
+  function setSelected(index: number) {
+    selectedIndex = index;
+  }
 </script>
 
 <List class="demo-list">
-  <ListItem text="Sent Mail">
-    <ListItemAdornment slot="leading-adornment" class="material-symbols-outlined">
-      send
-    </ListItemAdornment>
+  <ListItem selected={selectedIndex === 0} on:click={() => setSelected(0)}>
+    <Typography>Sent Mail</Typography>
+    <ListItemAdornment slot="leading" class="material-symbols-outlined">send</ListItemAdornment>
   </ListItem>
-  <ListItem text="Drafts">
-    <ListItemAdornment slot="leading-adornment" class="material-symbols-outlined">
-      drafts
-    </ListItemAdornment>
+  <ListItem selected={selectedIndex === 1} on:click={() => setSelected(1)}>
+    <Typography>Drafts</Typography>
+    <ListItemAdornment slot="leading" class="material-symbols-outlined">drafts</ListItemAdornment>
   </ListItem>
-
-  <ListItem text="Inbox" selected>
-    <ListItemAdornment slot="leading-adornment" class="material-symbols-outlined">
-      inbox
-    </ListItemAdornment>
+  <ListItem selected={selectedIndex === 2} on:click={() => setSelected(2)}>
+    <Typography>Inbox</Typography>
+    <ListItemAdornment slot="leading" class="material-symbols-outlined">inbox</ListItemAdornment>
   </ListItem>
 </List>