+++ website/src/routes/menu/menu-def-clickaway.svelte
@@ -1,0 +1,13 @@
+<script lang="ts">
+  import { Menu, Button, ListItem, Box, Typography } from '@uui';
+
+  let anchorRef: HTMLElement;
+  let open = false;
+</script>
+
+<Box ssx={{ $self: { display: 'flex', gap: '5rem', 'justify-content': 'center' } }}>
+  <Button variant="text" on:click={() => (open = !open)} bind:ref={anchorRef}>Clickaway</Button>
+  <Menu bind:open bind:anchorRef anchorOrigin="bottom-start">
+    <ListItem><Typography>List Item</Typography></ListItem>
+  </Menu>
+</Box>